services/AIService/requirements.txt
langchain==0.3.27
langchain-core==0.3.79
langchain-community==0.3.17
langchain_google_community==2.0.7
langchain-openai==0.2.1
langchain-groq==0.2.5
langchain-huggingface==0.3.1
langchain-google-genai==2.0.10
langchain_tavily
langgraph

google-generativeai>=0.7.0

sentence-transformers
faiss-cpu==1.7.3

fastapi
uvicorn
python-multipart
python-dotenv

pydub
requests
python-dateutil
pytz
pandas
PyPDF2

kafka-python
